08.11.16 Added range_car, made thumbnail to the left of article

+++ ftp/page-cart.php
@@ -49,7 +49,7 @@
 
 							//для умножения
 							settype($num, 'integer');
-							settype($price, 'integer');
+							settype($price, 'float');
 
 							$tovar_sum = $price * $num;
